snid: simplify unix backend, get rid of hostname directories
Instead of looking for the socket at $ROOT/$HOSTNAME/socket, snid
now looks for the socket at $ROOT/$HOSTNAME.

The rationale for using a dedicated directory per hostname was so
socket creation permissions could be set on a per-hostname basis.
However, managing the directories proved to be annoying in practice.
It's still possible to set permissions on a per-hostname basis
by symlinking from $ROOT/$HOSTNAME to a location in a directory
with the desired permissions.